PLFM-9752 - Fold benefactor filtering test into SearchIndexLifecycleWorkerAutowireTest
Merge the row-level benefactor filtering scenario into the existing
SearchIndexLifecycleWorkerAutowireTest instead of a separate file, matching
that test's entityService/entityAclManager conventions.
